% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aHull_function.R
\encoding{UTF-8}
\name{aHull}
\alias{aHull}
\title{aHull function}
\usage{
aHull (occ, crs, dist)
}
\arguments{
\item{occ}{Occurrences records of the species. It might be a 'sp.occ'
object corresponding to a list of 'SpatialPoints' for multiple species
(see \code{\link[aoh]{readOcc}} to obtain such class of object), a 'list' of
'data.frames' with the occurrences from multiple species, or a path for a folder
with the species occurrences files (.csv format). Each file, corresponding to
only one species and named with the corresponding species names, must have be 3 columns
identified as "species" (species names or other identification of taxa), "long"
(longitude), "lat" (latitude).}

\item{crs}{The Coordinate Reference System (CRS) specifing the projection and
datum of dataset. Could be a CRS object or a character string.}

\item{dist}{A value corresponding to the minimum distance assigned to consider
two coordinates as not duplicate. Values up to this distance will be consider
as duplicates and removed. Default is zero (i.e. only exactly coincindent
coordinates will be removed). Units of this value must be the same as those of
the coordinates for projected data or in km if coordinates are defined to be
longitude/latitude. For more details, see \code{\link[sp:remove.duplicates]
{remove.duplicates}}.}

\item{poly}{Optional. A polygon (ESRI shapefile as 'SpatialPolygonsDataFrame'
class) of the specific area to be checked.}

\item{fraction}{The minimum fraction of occurrences that must be included in polygon.}

\item{partCount}{The maximum number of disjunct polygons that are allowed.}
}
\value{
\code{aHull} returns the alpha hull polygon (ESRI shapefile as
'SpatialPolygonsDataFrame') representing the extent of occurrence (EOO) of
species and the respective alpha value assigned to build the provided polygon.
}
\description{
Provide the alpha hull polygon to represent the extent of occurrence (EOO) of
species from the occurrence records of multiple species. It is build based on
the original function \code{\link[rangeBuilder:getDynamicAlphaHull]
{getDynamicAlphaHull}} from rangeBuilder package, which allows that parameter
alpha be determined by the spatial distribution of the coordinates.
}
\author{
Thaís Dória & Daniel Gonçalves-Souza
}
