% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aHull_function.R
\encoding{UTF-8}
\name{aHull2}
\alias{aHull2}
\title{aHull function}
\usage{
aHull (occ, crs, dist = 0.25, poly = NULL, fraction = 1.0, partCount = 2,
buff = 1000, alphaIncrement = 0.01, cropToPoly = FALSE)
}
\arguments{
\item{occ}{Occurrences records of the species. It might be a 'sp.occ'
object corresponding to a list of 'SpatialPoints' for multiple species
(see \code{\link[aoh]{readOcc}} to obtain such class of object), a 'list' of
'data.frames' with the occurrences from multiple species, or a path for a folder
with the species occurrences files (.csv format). Each file, corresponding to
only one species and named with the corresponding species names, must have be 3 columns
identified as "species" (species names or other identification of taxa), "long"
(longitude), "lat" (latitude).}

\item{crs}{The Coordinate Reference System (CRS) specifing the projection and
datum of dataset. Could be a CRS object or a character string.}

\item{dist}{A value corresponding to the minimum distance assigned to consider
two coordinates as not duplicate. Values up to this distance will be consider
as duplicates and removed. Default is zero (i.e. only exactly coincindent
coordinates will be removed). Units of this value must be the same as those of
the coordinates for projected data or in km if coordinates are defined to be
longitude/latitude. For more details, see \code{\link[sp:remove.duplicates]
{remove.duplicates}}.}

\item{poly}{Optional. A polygon (ESRI shapefile as 'SpatialPolygonsDataFrame'
class) of the specific area to be checked for occurrences of species and to
restrict alpha hull development only to the species occurring inside of the polygon.
See details.}

\item{fraction}{The minimum fraction of occurrences that must be included in polygon.}

\item{partCount}{The maximum number of disjunct polygons that are allowed.}

\item{buff}{A buffer distance in meters to increase boundaries of alpha hull.}

\item{alphaIncrement}{The amount to increase alpha with each iteration.}

\item{cropToPoly}{(logical) Whether the output should also include the alpha
hull cropped by the provided poly. Only used if poly provided.}
}
\value{
\code{aHull} returns the alpha hull polygon (ESRI shapefile as
'SpatialPolygonsDataFrame') representing the extent of occurrence (EOO) of
species and the respective alpha value assigned to build the provided polygon.
If a \emph{poly} is provided and \emph{cropToPoly} is \code{TRUE}, \code{aHull}
also returns the cropped alpha hull.
#' @details The function generate an alpha hull polygon from a set of occurrences
records of multiple based on derivate ahull by sequentially increasing α parameter
(starting from 0 in steps of 0.01) until find the smallest value that returned one
single contiguous polygon encompassing all occurrences recordsincrease
If \emph{poly} is provided, the function filter the original species dataset by
keeping only those species occurring inside of the polygon. In this case, the
alpha hull construction is restrict to those filtered species, but the extend of
these polygons can extrapolate the area of polygon if are occurrences beyond the boundaries of 'poly' objetc. If the
spatial is desired, assign TRUE to 'clipToPoly'.
generated
, even if there are records
besides this area
}
\description{
From the raw occurrences records of multiple species, provides the alpha hull
polygon to represent the extent of occurrence (EOO) of each species which have
a minimun of 3 not duplicate records. It is build based on the original function
\code{\link[rangeBuilder:getDynamicAlphaHull]{getDynamicAlphaHull}}
(rangeBuilder package), that determines the parameter alpha by the spatial
distribution of the coordinates.
}
\seealso{
Dynamic Alpha hulls based on increase of alpha values are created with
}
\author{
Thaís Dória & Daniel Gonçalves-Souza
}
