% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ahull_plantRas}
\alias{ahull_plantRas}
\title{EOO raster for Plants (ahull)}
\format{
An 'aHull' object corresponding to a list with 5 RasterLayers.
}
\usage{
data(ahull_plantRas)
}
\description{
Dataset with RasterLayer exemplifying how the distribution file for the
argument \code{eooSp} of the function \code{aohMixS} should be structured.
}
\keyword{datasets}
