% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ahull_plantShp}
\alias{ahull_plantShp}
\title{EOO shapefile for Plants (ahull)}
\format{
An 'aHull' object corresponding to a list with 5 SpatialPolygons.
}
\usage{
data(ahull_plantShp)
}
\description{
Dataset with SpatialPolygons exemplifying how the distribution file for the
argument \code{eooSp} of the function \code{aohMixS} should be structured.
}
\keyword{datasets}
