% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aHull_function.R
\encoding{UTF-8}
\name{aHullt2}
\alias{aHullt2}
\title{aHull: generate alpha hull polygons for multiple-species}
\usage{
aHull (occ, crs, alphaIncrement = 0.01, fraction = 1.0, partCount = 2, buff = 1000, distOcc = 0.25, poly = NULL, cropToPoly = FALSE)
}
\arguments{
\item{occ}{Occurrences records of the species. It might be a 'spOcc'
object corresponding to a list of 'SpatialPoints' for multiple species
(see \code{\link[aoh]{readOcc}} to obtain such class of object), a 'list' of
'data.frames' with the occurrences from multiple species, or a path for a folder
with the species occurrences files (.csv format). Each file, corresponding to
only one species and named with the corresponding species names, must have be 3 columns
identified as "species" (species names or other identification of taxa), "long"
(longitude), "lat" (latitude). NOTE: Longitude must be at a column before the
latitude column.}

\item{crs}{The Coordinate Reference System (CRS) specifing the projection and
datum of dataset. Could be a CRS object or a character string.}

\item{fraction}{The minimum fraction of occurrences that must be included in polygon.}

\item{partCount}{The maximum number of disjunct polygons that are allowed.}

\item{alphaIncrement}{The amount to increase alpha with each iteration.}

\item{buff}{A buffer distance in meters to increase boundaries of alpha hull.}

\item{distOcc}{A value corresponding to the minimum distance assigned to consider
two coordinates as not duplicate. Values up to this distance will be consider
as duplicates and removed. Default is zero (i.e. only exactly coincindent
coordinates will be removed). Units of this value must be the same as those of
the coordinates for projected data or in km if coordinates are defined to be
longitude/latitude. For more details, see \code{\link[sp:remove.duplicates]
{remove.duplicates}}.}

\item{poly}{Optional. A polygon (ESRI shapefile as 'SpatialPolygonsDataFrame'
class) of the specific area to be checked for occurrences of species and to
restrict alpha hull development only to the species occurring inside of the polygon.
See details.}

\item{cropToPoly}{(logical) Whether the output should also include the alpha
hull cropped by the provided poly. Only used if poly provided.}
}
\value{
\code{aHull} returns the alpha hull polygon (ESRI shapefile as
'SpatialPolygonsDataFrame') representing the extent of occurrence (EOO) of
species and the respective alpha value assigned to build the alpha hull.
If a \emph{poly} is provided and \emph{cropToPoly} is \code{TRUE}, \code{aHull}
also returns the cropped alpha hull.
}
\description{
From the raw occurrences records of multiple species, provides the alpha hull
polygon representing the extent of occurrence (EOO) of each species which have
a minimun of 3 not duplicate records. It is build based on the original function
\code{\link[rangeBuilder]{getDynamicAlphaHull}} (rangeBuilder package), that
determines the α parameter by the spatial distribution of the coordinates.
}
\details{
The function generate an alpha hull polygon from the occurrences
records of multiple species by sequentially increasing α parameter (starting from 0
in steps of defined \emph{alphaIncrease} until find the smallest value that return the
\emph{partCount} polygon encompassing the \emph{fraction} of occurrences provided.
If \emph{poly} is provided, the function filter the original species dataset by
keeping only those species occurring inside of the specified polygon. In this case, the
alpha hull construction is restrict to the filtered species, but the extension of
these polygons can extrapolate the area of polygon if there are occurrences beyond
the boundaries of 'poly' object. If the user want also have the alpha hull cropped
by the area of \emph{poly}, assign \code{TRUE} to 'cropToPoly'.
}
\seealso{
Dynamic alpha hulls from increasing alpha values are created with
\code{\link[rangeBuilder:getDynamicAlphaHull]{getDynamicAlphaHull}}
}
\author{
Thaís Dória & Daniel Gonçalves-Souza
}
