% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{al}
\alias{al}
\title{Altitude Map}
\format{A RasterLayer with 2395 rows, 1681 columns and resolution of
0.005 x 0.005 degrees}
\usage{
data(al)
}
\description{
A fictitious dataset exemplifying how the raster for the argument \code{alt.map}
of the function \code{aoh} should be structured.
}
\keyword{datasets}
