% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{alpref}
\alias{alpref}
\title{Altitude preference for Amphibians}
\format{A data frame with 99 rows and 3 variables}
\usage{
data(alpref)
}
\description{
A fictitious dataset exemplifying how the data frame for the argument 'matrix.alt.pref'
of the function aoh should be structured.
}
\keyword{datasets}
