% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{alt_map}
\alias{alt_map}
\title{Altitude Map}
\format{A RasterLayer with 266 rows, 187 columns and resolution of
0.05 x 0.05 degrees}
\usage{
data(alt_map)
}
\description{
A fictitious dataset exemplifying how the raster for the argument \code{alt.map}
of the function \code{aoh} should be structured.
}
\keyword{datasets}
