% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoh_function.R
\name{aoh}
\alias{aoh}
\title{AOH function}
\usage{
aoh(sd, lc.rec, matrix.hab.pref, alt.map = NULL, matrix.alt.pref,
  shp.out = FALSE)
}
\arguments{
\item{sd}{Spatial distribution shapefile. The name of the species must be on the second column of tbe attribute table of the shapefile.}

\item{lc.rec}{Land use map reclassified for the categories of habitat}

\item{matrix.hab.pref}{Data frame 0/1 of habitat preference of the species. First #column must be the species name. The posterior columns must be named after the categories of habitat as folowing the lc.rec classification}

\item{alt.map}{Elevation map}

\item{matrix.alt.pref}{Data frame with altitudinal range of species. First column must be the species name, second column the min value of altitude and the third column the max value of altitude}

\item{shp.out}{(logical) Whether the output should be a shapefile as opposed to a raster}
}
\value{
The result is a RasterLayer or RasterBrick; or SpatialPolygons object
}
\description{
Provide the area of habitat (AOH) of a given species through refinement of its known geographic distribution
}
\details{
The function map the area of habitat within the geographical distribution (SpatialPolygon) given as the input data. as the refined distribution of a given species. This refinement is made considering the specific preference for habitats of a given species.
}
\author{
Daniel Gonçalves-Souza & Thaís Dória
}
