% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{aoh_amphRas}
\alias{aoh_amphRas}
\title{Spatial distribution corresponding to 'Area of Habitat' (AOH) for Amphibians as
raster file format (.asc).}
\format{A RasterStack with 3 layers.}
\usage{
data(aoh_amphRas)
}
\description{
A fictitious dataset exemplifying how the raster file for the argument \code{aoh}
of the function \code{aohVal} could be structured.
}
\keyword{datasets}
