% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{aoh_amphShp}
\alias{aoh_amphShp}
\title{Spatial distribution corresponding to the 'Area of Habitat' (AOH) for Amphibians as
SpatialPolygons in ESRI shapefile (.shp)}
\format{A SpatialPolygons with 3 features.}
\usage{
data(aoh_amphShp)
}
\description{
A fictitious dataset exemplifying how the spatial distribution file for
the argument \code{aoh} of the function \code{aohVal} could be structured.
}
\keyword{datasets}
