% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{aoh_birdRas}
\alias{aoh_birdRas}
\title{AOH raster for Birds}
\format{
A RasterStack with 3 layers.
}
\usage{
data(aoh_birdRas)
}
\description{
Spatial distribution corresponding to 'Area of Habitat' (AOH) for Birds as
raster file format (.asc).
A fictitious dataset exemplifying how the raster file for the argument \code{aoh}
of the function \code{aohVal} could be structured.
}
\keyword{datasets}
