% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{aoh_birdShp}
\alias{aoh_birdShp}
\title{AOH shapefile for Birds}
\format{
A SpatialPolygons with 4 features.
}
\usage{
data(aoh_birdShp)
}
\description{
Spatial distribution corresponding to the 'Area of Habitat' (AOH) for Birds as
SpatialPolygons in ESRI shapefile (.shp)
A fictitious dataset exemplifying how the spatial distribution file for
the argument \code{aoh} of the function \code{aohVal} could be structured.
}
\keyword{datasets}
