% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aohMix_function.R
\encoding{UTF-8}
\name{aohMix}
\alias{aohMix}
\title{aohMix: mapping the AOH of species by mixing data from SDM and EOO following Syfert et al. (2014).}
\usage{
aohMix1 (eooSp, modSp, thresInitial = NULL , thresIncrement = NULL,
continuous = TRUE, poly = NULL, cropToPoly = FALSE, progress = TRUE,
removeTempFile = TRUE, stack = TRUE )
}
\arguments{
\item{eooSp}{Spatial distribution data of the species representing the
original (i.e. not refined) extent of occurrence (EOO). It might correspond to
\itemize{
  \item path for a folder with spatial distribution files (.shp, .asc or .tif
  format)
  \item SpatialPolygonDataFrame (see \code{\link[aoh]{readShp}} to obtain such
class of object)
  \item aHull object created from occurrences records (see \code{\link[aoh]{aHull}}
  to obtain such class of object)
  \item 'aHull' Raster
  \item 'RasterLayer' object
  \item 'RasterStack' object
  \item 'RasterBrick' object
 }}

\item{thresInitial}{teste}

\item{thresIncrement}{teste}
}
\description{
Based on a mixing approach to combine the geographical species distribution
derived from spatial-informed methods and also modelling techniques, mapping the specie's
area of habitat (AOH) from a mixing approach
}
\details{
mapping the AOH by thresholding distribution models (SDM) based on the EOO geometry.
}
\examples{

### Shapefile as Input ('aHull' class) ###

# Binary Output #

aohmixS_bin <- aohMix (eooSp = aHull_plantShp, modSp = sdm_plantRas,
thresInitial = 0.05, thresIncrement = 0.25, continuous = FALSE, cropToPoly = TRUE,
poly = poly, progress = TRUE, stack = TRUE)

# Continuous Output #

aohmixS_con <- aohMix (eooSp = aHull_plantShp, modSp = sdm_plantRas,
thresInitial = 0.05, thresIncrement = 0.25, continuous = TRUE, cropToPoly = TRUE,
poly = poly, progress = TRUE, stack = TRUE)


### Raster as Input ###

# Binary Output #

aohmixR_bin <- aohMix (eooSp = aHull_plantRas, modSp = sdm_plantRas,
thresInitial = 0.05, thresIncrement = 0.25, continuous = FALSE, cropToPoly = TRUE,
poly = poly, progress = TRUE, stack = TRUE)

# Continuous Output #

aohmixR_con <- aohMix (eooSp = aHull_plantRas, modSp = sdm_plantRas,
thresInitial = 0.05, thresIncrement = 0.25, continuous = TRUE, cropToPoly = TRUE,
poly = poly, progress = TRUE, stack = TRUE)

}
\references{
1. Syfert M.M., Joppa L.N., Smith M.J, Coomes D., Bachman S.P. & Brummitt N.A. (2014).
Using species distribution models to inform IUCN Red List assessments. Biological
Conservation. 177.

2. Sangermano F., Eastman J.R. (2012). A GIS framework for the refinement of species
geographic ranges. International Journal of Geographical Information Science 26, 39-55.
}
\author{
Thaís Dória & Daniel Gonçalves-Souza
}
