% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aohMixS_function.R
\encoding{UTF-8}
\name{aohMixS}
\alias{aohMixS}
\title{aohMixS: mapping the species' AOH oby mixing data from SDM and EOO following
Syfert et al. (2014).}
\usage{
aohMixS (eooSp = eooSp, modSp = modSp, thresInitial = NULL , thresIncrement = NULL,
continuous = TRUE, poly = NULL, cropToPoly = FALSE, progress = TRUE,
removeTempFile = TRUE, stack = TRUE )
}
\arguments{
\item{eooSp}{Spatial distribution data of the species representing the
original (i.e. not refined) extent of occurrence (EOO). It might correspond to
\itemize{
  \item path for a folder with spatial distribution files (.shp, .asc or .tif
  format)
  \item SpatialPolygonDataFrame (see \code{\link[aoh]{readShp}} to obtain such
class of object)
  \item aHull object created from occurrences records (see \code{\link[aoh]{aHull}}
  to obtain such class of object)
  \item 'aHull' Raster
  \item 'RasterLayer' object
  \item 'RasterStack' object
  \item 'RasterBrick' object
 }}

\item{thresIncrement}{The amount to increase threshold with each iteration.}

\item{continuous}{(logical) Whether the output should be binary or continuous
for the rasters.}

\item{poly}{Optional. A polygon (ESRI shapefile in a 'SpatialPolygonsDataFrame'
class) of a given area to be used as a mask to crop and restrict the models-derived
EOOs to the area corresponding to provided polygon. NOTE: this argument does not
reduce the set of occurrences based on the poly, but only reduce the species dataset based on its presence inside of poly.
See details.}

\item{cropToPoly}{(logical) Whether the output should also include the alpha
hull cropped by the provided poly. Only used if 'poly' is provided. Default is
\code{FALSE}.}

\item{stack}{(logical) Whether the output should be returned as a stack of 'RasterLayer'
files. Default is \code{TRUE}.}
}
\description{
Through a mix approach that combines the species' spatial-informed distribution
with modelling techniques, thresholds the distribution models (SDM) based on EOO's
geometry to generate models-derived EOOs. It is develop by adapting the approach
described in Syfert et al. (2014), so that the species' area of habitat (AOH) can
be interpreted as the SDMs-derived Alpha Hull (see Syfert et al. 2014).
}
\examples{

### Shapefile as Input ('aHull' class) ###

# Binary Output #

aohmixS_bin <- aohMixS (eooSp = aHull_plantShp, modSp = sdm_plantRas,
thresInitial = 0.05, thresIncrement = 0.25, continuous = FALSE, cropToPoly = TRUE,
poly = poly, progress = TRUE, stack = TRUE)

# Continuous Output #

aohmixS_con <- aohMixS (eooSp = aHull_plantShp, modSp = sdm_plantRas,
thresInitial = 0.05, thresIncrement = 0.25, continuous = TRUE, cropToPoly = TRUE,
poly = poly, progress = TRUE, stack = TRUE)


### Raster as Input ###

# Binary Output #

aohmixR_bin <- aohMixS (eooSp = aHull_plantRas, modSp = sdm_plantRas,
thresInitial = 0.05, thresIncrement = 0.25, continuous = FALSE, cropToPoly = TRUE,
poly = poly, progress = TRUE, stack = TRUE)

# Continuous Output #

aohmixR_con <- aohMixS (eooSp = aHull_plantRas, modSp = sdm_plantRas,
thresInitial = 0.05, thresIncrement = 0.25, continuous = TRUE, cropToPoly = TRUE,
poly = poly, progress = TRUE, stack = TRUE)

}
\references{
1. Syfert M.M., Joppa L.N., Smith M.J, Coomes D., Bachman S.P. & Brummitt N.A. (2014).
Using species distribution models to inform IUCN Red List assessments. Biological
Conservation. 177.

2. Sangermano F., Eastman J.R. (2012). A GIS framework for the refinement of species
geographic ranges. International Journal of Geographical Information Science 26, 39-55.
}
\author{
Thaís Dória & Daniel Gonçalves-Souza
}
