% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkOcc_function.R
\encoding{UTF-8}
\name{checkOcc}
\alias{checkOcc}
\title{checkOcc: check if occurrences fall inside of an specific are}
\usage{
checkOcc (occ, poly, distOcc = NULL, SpOcc = FALSE)
}
\arguments{
\item{occ}{Occurrences records of the species (coordinates in decimal degrees).
It might be a path for a folder with the species occurrences files (.csv format),
a 'list' of 'data.frames' with the occurrences from multiple species (see data
examples), or a 'spOcc' object corresponding to a list of 'SpatialPoints' from
multiple species (see \code{\link[aoh]{readOcc}} to obtain such object).
NOTE: If path is provided, each .csv file should correspond to only one species
and the file must be named with the corresponding species names. The files must
have 3 columns identified as "species" (species names or other identification of taxa),
"long" (longitude), "lat" (latitude). Longitude must be in a column before the
latitude column.}

\item{poly}{A polygon (ESRI shapefile from a 'SpatialPolygonsDataFrame' class)
of the specific area to be checked.
#' @param distOcc A value corresponding to the minimum distance assigned to
consider two coordinates as not duplicate. Values up to this distance will
correspond to duplicates and removed. Units of this value must be in km. Default
is zero (i.e. only exactly coincindent coordinates will be removed).
For more details,see \code{\link[sp:remove.duplicates]{remove.duplicates}}.
Optional and only used if 'occ' is a path for .csv files or a list of 'data.frame'.
If 'occ' correspond to 'SpatialPoints', this argument ('distOcc') should be ignored
(i.e. NULL).}

\item{SpOcc}{(logical) Whether the output should also return the 'SpatialPoints'
of species filtered based on the poly boundaries. Default is \code{FALSE}.}
}
\value{
\code{checkOcc} returns a data.frame of species and the respectives
occurrences matching with the provided polygon. If SpOcc is \code{TRUE},
\code{checkdOcc} returns a list with two elements. The first element is the
data.frame with the species and occurrences filtered by the polygon. The second
is the 'SpatialPoints' of each species that match with polygon. In this case,
the 'poly' is used only to reduce the species from initial dataset, but not
to restrict the 'SpatialPoints' to occurrences inside of polygon. 'SpatialPoints'
returned, thus, correspond to all set of occurrences and not to a subset that matched
with the provided polygon (i.e. it is not cleaned to remove the records
falling outside the polygon checked).
}
\description{
From a list of 'SpatialPoints' for multiple species, gives the species and its
occurrences that are inside of a polygon (ESRI shapefile format) representing
a region of interest (e.g. study area, continent, country or other regions with
specific boundaries).
}
\author{
Thaís Dória & Daniel Gonçalves-Souza
}
