% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{climSuit_bird}
\alias{climSuit_bird}
\title{Climatic suitability map for Birds}
\format{A list with 3 Rasterlayers.}
\usage{
data(climSuit_bird)
}
\description{
A fictitious dataset exemplifying how the list file for the argument \code{climSuit}
of the function \code{aoh} should be structured.
}
\keyword{datasets}
