% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{df_plantOcc5spp}
\alias{df_plantOcc5spp}
\title{Occurrence records for Plants}
\format{
A list with 5 data frames.
}
\usage{
data(df_plantOcc5spp)
}
\description{
Dataset with geographical coordinates exemplifying how the occurrences data file
for the argument \code{occ} of the function \code{aohMixK} should be structured.
}
\keyword{datasets}
