% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{eoo_amphRas}
\alias{eoo_amphRas}
\title{Spatial distribution corresponding to original (i.e. nof refined) 'Extent of Occurrence' (EOO)
for Amphibians as raster file format (.asc)}
\format{A RasterStack with 3 layers.}
\usage{
data(eoo_amphRas)
}
\description{
A fictitious dataset exemplifying how the raster file for the argument \code{eoo}
of the function \code{aohVal} could be structured.
}
\keyword{datasets}
