% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{eoo_amphShp}
\alias{eoo_amphShp}
\title{Spatial distribution corresponding to original (i.e. nof refined) 'Extent of Occurrence' (EOO)
for Amphibians as SpatialPolygons in ESRI shapefile (.shp)}
\format{A SpatialPolygons with 3 features.}
\usage{
data(eoo_amphShp)
}
\description{
A fictitious dataset exemplifying how the raster file for the argument \code{eoo}
of the function \code{aohVal} could be structured.
}
\keyword{datasets}
