% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{eoo_birdRas}
\alias{eoo_birdRas}
\title{EOO Raster for Birds}
\format{
A RasterStack with 3 layers.
}
\usage{
data(eoo_birdRas)
}
\description{
Extent of Occurrence' (EOO) corresponding to the original (i.e. not refined)
 spatial distribution of Birds in class of 'SpatialPolygonsDataFrame'.
A fictitious dataset exemplifying how the raster file for the argument \code{eoo.sp}
of the function \code{aohVal} could be structured. This dataset was generated
based on available data from the IUCN database. To provide an example dataset
for this package, the original maps were thinned and lightly custom modified.
}
\keyword{datasets}
