% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{habpref}
\alias{habpref}
\title{Habitat preference for Amphibians}
\format{A data frame with 129 rows and 5 variables}
\usage{
data(habpref)
}
\description{
A fictitious dataset exemplifying how the data frame for the argument \code{matrix.hab.pref}
of the function \code{aoh} should be structured.
}
\keyword{datasets}
