% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{habpref_bird}
\alias{habpref_bird}
\title{Habitat preference for Birds}
\format{A data frame with 4 rows and 4 variables}
\usage{
data(habpref_bird)
}
\description{
A fictitious dataset exemplifying how the data frame for the argument \code{habPref}
of the function \code{aoh} should be structured.
}
\keyword{datasets}
