% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{lc}
\alias{lc}
\title{Land Cover}
\format{A RasterLayer with 4790 rows, 3361 columns and resolution of
0.0028 x 0.0028 degrees}
\usage{
data(lc)
}
\description{
A fictitious dataset exemplifying how the raster for the argument 'lc.rec'
of the function aoh should be structured.
}
\keyword{datasets}
