% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{lc_map}
\alias{lc_map}
\title{Land Cover}
\format{A RasterLayer with 1198 rows, 841 columns and resolution of
0.01 x 0.01 degrees}
\usage{
data(lc_map)
}
\description{
A fictitious dataset exemplifying how the raster for the argument \code{lc}
of the function \code{aoh} should be structured.
}
\keyword{datasets}
