% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{plants_thresholds}
\alias{plants_thresholds}
\title{Thresholds for Plants' SDMs}
\format{
A data frame with 5 rows and 1 variables
}
\usage{
data(plants_thresholds)
}
\description{
A fictitious dataset exemplifying how the data frame for the argument
\code{threshold} of the function \code{aohMixK} should be structured.
}
\keyword{datasets}
