% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aohVal_plot.R
\encoding{UTF-8}
\name{plot.aohVal}
\alias{plot.aohVal}
\title{Plot 'aohVal' objects}
\usage{
plot(aohVal)
}
\arguments{
\item{aohVal}{A object from a class 'aohVal', which can be obtained with
\code{\link[habitaR]{aohVal}} function.}
}
\value{
\code{plot.aohVal} returns a graphical representation of 'validation
analysis performed with \code{\link[habitaR]{aohVal}} following Rondinini
et al. (2011).
}
\description{
Provide a \code{\link[graphisc]{plot}} of the graphical representation of
validation results from the species AOH mapping.
}
\author{
Thaís Dória & Daniel Gonçalves-Souza
}
