% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{poly}
\alias{poly}
\title{Shapefile of Caatinga}
\format{
A SpatialPolygonsDataFrame object.
}
\usage{
data(poly)
}
\description{
Polygon of an region exemplifying how the polygon file for the
argument \code{poly} of the functions \code{checkOcc}, \code{aHull} and
\code{aohMixS} should be structured.
}
\keyword{datasets}
