% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pref_hab_function.R
\name{prefHab}
\alias{prefHab}
\title{Habitat preference}
\usage{
prefHab(sp, key, cat = NULL, progress = F)
}
\arguments{
\item{progress}{}
}
\value{
The result is a list with two elements. The first element is a
data.frame detailing if the function was able (1) or not (0) to refinate the
species distribution. The second element is a list of RasterLayer or
SpatialPolygons object representing the refinated distribution of the species.
For the RasterLayer the value of 0 indicates the cells where the the species
were considerated present before the refinement. Values higher than 0 indicates
the coverage of cell by the spatial distribution after the refinement.
}
\description{
Provide the area of habitat (AOH) of a given species through refinement of its
known geographic distribution
}
\details{
The function map the area of habitat within the geographical
distribution (SpatialPolygon) given as the input data. as the refined
distribution of a given species. This refinement is made considering the
specific preference for habitats of a given species.
}
\examples{
data(sd_amph)
data(lc)
data(habpref)
data(al)
data(alpref)
ref_data <- aoh(eoo = sd_amph, lc.rec = lc , matrix.hab.pref = habpref,
alt.map = al, matrix.alt.pref = alpref)
}
\author{
Daniel Gonçalves-Souza & Thaís Dória
}
