% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ras}
\alias{ras}
\title{Raster model}
\format{
A RasterLayer.
}
\usage{
data(ras)
}
\description{
A fictitious dataset exemplifying how the raster file for the
argument \code{ras} of the function \code{aHull} should be structured.
}
\keyword{datasets}
