% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOcc_function.R
\encoding{UTF-8}
\name{readOcc}
\alias{readOcc}
\title{readOcc: read species occurrences}
\usage{
readOcc (occ, crs = "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0",
distOcc = 0.25, occSum = FALSE)
}
\arguments{
\item{occ}{Path for a folder with the species occurrences records files
(.csv format). Coordinates should be in decimal degrees. The names of files
should correspond to the species names and must have be 3 columns identified as
"species" (species names or other identification of taxa), "long" (longitude),
"lat" (latitude). NOTE: Longitude must be at a column before the latitude column.}

\item{crs}{The Coordinate Reference System (CRS) specifing the projection and
datum of dataset. Could be a CRS object or a character string.}

\item{distOcc}{A value corresponding to the minimum distance assigned to consider
two coordinates as not duplicate. Values up to this distance will be consider
as duplicates and removed. Units of this value must be in km. Optional.
Default is zero (i.e. only exactly coincindent coordinates will be removed).
For more details, see \code{\link[sp:remove.duplicates]{remove.duplicates}}.}

\item{OccSum}{(logical) Whether the output should include also a data.frame
with the number of occurrences records after the removal of duplicate coordinates.
Default if \code{FALSE}.}
}
\value{
By default, \code{readOcc} returns a 'spOcc' object corresponding to a
species list with features from a class of 'SpatialPoints'.
If occSum is \code{TRUE}, \code{readOcc} also returns a data.frame with the number
of occurrences records after the cleaning step to remove duplicate coordinates.
}
\description{
Read and clean the occurrences records files (.csv format) of multiples species
from a specified folder and provide a list of 'SpatialPoints' of species through
the removal of duplicates coordinates.
}
\seealso{
\code{\link[sp:remove.duplicates]{remove.duplicates}}
}
\author{
Thaís Dória & Daniel Gonçalves-Souza
}
