% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readShp_funciton.R, R/readShp_function.R
\name{readShp}
\alias{readShp}
\title{readSHP function}
\usage{
readShp (path)

readShp (path)
}
\arguments{
\item{path}{Path for a folder with spatial distribution shapefiles
(ESRI shapefile format). The name of the species must be on the second
column of the attribute table of the shapefile.}

\item{path}{Path for a folder with spatial distribution shapefiles
(ESRI shapefile format). The name of the species must be on the second
column of the attribute table of the shapefile.}
}
\value{
A object from a class of 'SpatialPolygonDataFrame'

A object from a class of 'SpatialPolygonDataFrame'
}
\description{
Read the spatial distribution shapefiles (ESRI shapefile format) of species
from a specified folder and provide a SpatialPolygonDataFrame object.

Read the spatial distribution shapefiles (ESRI shapefile format) of species
from a specified folder and provide a SpatialPolygonDataFrame object.
}
\author{
Thaís Dória & Daniel Gonçalves-Souza

Thaís Dória & Daniel Gonçalves-Souza
}
