% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sd_amph}
\alias{sd_amph}
\title{Spatial distribution for Amphibians Map}
\format{A SpatialPolygonsDataFrame 4 features.}
\usage{
data(sd_amph)
}
\description{
A fictitious dataset exemplifying how the shapefile for the argument 'eoo'
of the function aoh could be structured.
}
\keyword{datasets}
