% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sdm_plantRas}
\alias{sdm_plantRas}
\title{Lorem ipsum}
\format{
A list with 6 Rasterlayers.
}
\usage{
data(sdm_plantRas)
}
\description{
Lorem ipsum Lorem ipsum Lorem ipsum
}
\keyword{datasets}
