% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sdm_plantRas5spps}
\alias{sdm_plantRas5spps}
\title{SDM raster for Plants}
\format{
A list with 5 RasterLayers.
}
\usage{
data(sdm_plantRas5spp)
}
\description{
Dataset with RasterLayers exemplifying how the distribution model file for
the argument \code{modSp} of the functions \code{aohMixS} and \code{aohMixK}
should be structured.
}
\keyword{datasets}
