% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sdm_plantSta5spp}
\alias{sdm_plantSta5spp}
\title{SDMs stack for Plants}
\format{
A RasterStack with 5 RasterLayers.
}
\usage{
data(sdm_plantSta5spp)
}
\description{
Dataset with RasterStack exemplifying how the distribution model file for
the argument \code{modSp} of the functions \code{aohMixS} and \code{aohMixK}
should be structured.
}
\keyword{datasets}
