% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Sp_plantOcc5spp}
\alias{Sp_plantOcc5spp}
\title{Occurrence records for Plants}
\format{
A spOcc' object corresponding to a list with 5 SpatialPoints.
}
\usage{
data(Sp_plantOcc5spp)
}
\description{
Dataset with SpatialPoints exemplifying how the occurrences data file for the
argument \code{occ} of the function \code{aohMixK} should be structured.
}
\keyword{datasets}
