% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{spdf_plantShp5spp}
\alias{spdf_plantShp5spp}
\title{EOO shapefile for Plants}
\format{
A SpatialPolygonsDataFrame with 5 Polygons.
}
\usage{
data(spdf_plantShp5spp)
}
\description{
Dataset with SpatialPolygonsDataFrame exemplifying how the distribution file
for the argument \code{eooSp} of the function \code{aohMixS} should be
structured.
}
\keyword{datasets}
