% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{spOcc_plants}
\alias{spOcc_plants}
\title{Occurrence records for Plants}
\format{
A spOcc' object corresponding to a list with 12 SpatialPoints.
}
\usage{
data(spOcc_plants)
}
\description{
Dataset with SpatialPoints exemplifying how the occurrences data file for the
argument \code{occ} of the function \code{aHull} should be structured.
}
\keyword{datasets}
