% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spRecords_function.R
\encoding{UTF-8}
\name{spRecords}
\alias{spRecords}
\title{spRecords}
\usage{
spRecords(sp, ex = NULL, dataBase = c('GBIF', 'VertNet', 'BISON'),
progress = FALSE)
}
\arguments{
\item{sp}{a vector containing the names of the species.}

\item{ex}{The extent of the output Spatial points. Should be an extent object
or a list containing extent objects.}

\item{dataBase}{vector cortaining which databases the function should search
for occurrence data.}
}
\value{
\code{spRecords} returns a list containing SpatialPoints objects of
the occurrence of the species.
}
\description{
Gets occurrence data of species from databases and turn into a
SpatialPoints object#'
}
\examples{
name <- c('Orcinus orca', 'Bothrops jararaca')
area <- list(extent(c(-179.8733, 79.3333, -8.6333, 8.65083 )), NULL)
occ.shps <- spRecords(sp = name, ex = area, dataBase = c('GBIF', 'BISON'))

}
\author{
Daniel Gonçalves-Souza & Thaís Dória
}
