% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aHull_function.R
\encoding{UTF-8}
\name{aHull}
\alias{aHull}
\title{aHull: generate alpha hull polygons for multiple-species}
\usage{
aHull (occ, crs, fraction = NULL, partCount = NULL, alphaIncrement = NULL,
buff = 0.0, distOcc = NULL, poly = NULL, cropToPoly = FALSE, occSum = TRUE)
}
\arguments{
\item{occ}{Occurrences records of the species (coordinates in decimal degrees).
It might be a path for a folder with the species occurrences files (.csv format),
a 'list' of 'data.frames' with the occurrences from multiple species (see data
examples), or a 'spOcc' object corresponding to a list of 'SpatialPoints' from
multiple species (see \code{\link[aoh]{readOcc}} to obtain such object).
NOTE: If path is provided, each .csv file should correspond to only one species
and the file must be named with the corresponding species names. The files must
have 3 columns identified as "species" (species names or other identification of taxa),
"long" (longitude), "lat" (latitude). Longitude must be in a column before the
latitude column.}

\item{crs}{The Coordinate Reference System (CRS) specifing the projection and
datum of dataset. Could be a CRS object or a character string.}

\item{fraction}{The minimum fraction of occurrences that must be included in polygon.}

\item{partCount}{The maximum number of disjunct polygons that are allowed.}

\item{alphaIncrement}{The amount to increase alpha with each iteration.}

\item{buff}{A buffer distance in meters to increase boundaries of alpha hull.
Default is zero (0).}

\item{distOcc}{A value corresponding to the minimum distance assigned to
consider two coordinates as not duplicate. Values up to this distance will
correspond to duplicates and removed. Units of this value must be in km.
Default is zero (i.e. only exactly coincindent coordinates will be removed).
Optional and only used if 'occ' is a path for .csv files or a list of data.frames.
If 'occ' correspond to 'SpatialPoints', this argument should be ignored
(i.e., distOcc = NULL). For more details, see \code{\link[sp:remove.duplicates]{remove.duplicates}}
in the \pkg{sp} package.}

\item{poly}{Optional. A polygon (ESRI shapefile in a 'SpatialPolygonsDataFrame'
class) of a given area to be checked for occurrences of species and to restrict
the building of alpha hull only to the species occurring inside of the provided
polygon. See details.}

\item{cropToPoly}{(logical) Whether the output should also include the alpha
hull cropped by the provided poly. Only used if 'poly' is provided.}

\item{occSum}{(logical) Whether the output should include also a data.frame
with the number of occurrences records after the removal of duplicate coordinates.
Default is \code{TRUE}.
#'}
}
\value{
\code{aHull} returns a list with two elements. The first is a data.frame
of species and the respective alpha values and occurrences used to construct
the alpha hull. The second element is a 'aHull' object corresponding to a list
of alpha hulls of each species ('SpatialPolygons class). If the conditions
assigned by the user to build the alpha hulls cannot be satisfied, is returned
the minimum convex (MCH). If a \emph{poly} is provided and \emph{cropToPoly}
is \code{TRUE}, \code{aHull} also returns the cropped alpha hulls as a third
element from the resulting list.
}
\description{
From a set of occurrences records, provides the alpha hull representing the
extent of occurrence (EOO) of multiple species which have a minimun of 3 not
duplicate records.
It is build based on the original function \code{\link[rangeBuilder]{getDynamicAlphaHull}}
from the \pkg{rangeBuilder} package, that determines the α parameter by the spatial distribution of the
coordinates.
}
\details{
Based on a set of occurrences records of multiple-species, the
function generate alpha hull polygons ('SpatialPolygons' class) by sequentially
increasing the value of α parameter (starting from 0 in steps of defined
\emph{alphaIncrement}) until find the smallest value that met the \emph{partCount} condition and
encompass the \emph{fraction} of occurrences assigned by the user.
If \emph{poly} is provided, the function filter the original species dataset
by keeping only those species occurring inside of the specified region.
In this case, the construction of alpha hulls will be restricted only to the
filtered species, but the extension of these resulting maps can extrapolate
the area of 'poly' if the occurrences extend far away beyond the boundaries
of 'poly' object. NOTE: if the user want have the alpha hull cropped by the
area of \emph{poly}, 'cropToPoly' should be assign as \code{TRUE}, but even in
this case, the alpha hull is firstly constructed based on all dataset of
occurrences and then cropped by some desired region in a second step. Thus,
cropped alpha hulls correspond to the original alpha hull clipped/trinned by
some region and not drawn based in a restrict set of occurrences recorded inside
of some specific region.
}
\examples{

### Fictitious plants data

# Example for signature 'DataFrame' (occ).

ahull_plants<-aHull(occ = occ_plants, crs=, plot = TRUE, progress = TRUE)

}
\references{
1. Capinha Brooks, T. M, Fonseca, S.L. Pimm, Akçakaya, H.R., Buchanan, G.M., …,
Rondinini C. (2019). Measuring Terrestrial Area of Habitat (AOH) and Its
Utility for the IUCN Red List. Trends in Ecology &amp; Evolution, 34(11),
977–986.

2. CRIADOR DO RANGE BUILDER.

3. Meyer et al. (2018).
}
\seealso{
The dynamic alpha hulls drawn from the sequentially increase of alpha
values are created with \code{\link[rangeBuilder:getDynamicAlphaHull]{getDynamicAlphaHull}}
from \pkg{rangeBuilder} package.
}
\author{
Thaís Dória & Daniel Gonçalves-Souza
}
