% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoh_function.R
\name{aoh}
\alias{aoh}
\title{AOH function}
\usage{
aoh(eoo, lc.rec, matrix.hab.pref, alt.map = NULL, matrix.alt.pref = NULL,
shp.out = FALSE, resolution = NULL, continuous = FALSE, threshold = 0.5,
extent.out = NULL, progress = FALSE)
}
\arguments{
\item{eoo}{SpatialPolygons of the spatial distribution of the species or path
for a folder with spatial distribution shapefiles (ESRI shapefile format) .
The name of the species must be on the second column of the attribute table
of the shapefile.}

\item{lc.rec}{RasterLayer object of the land use map reclassified for the
categories of habitat.}

\item{matrix.hab.pref}{Data frame 0/1 of habitat preference of the species.
First column must be the species name. The posterior columns must be named
after the categories of habitat as following the lc.rec classification.
\code{\link[habitaR]{prefHab}} can help obtain this data.}

\item{alt.map}{RasterLayer object of the elevation map. Optional.}

\item{matrix.alt.pref}{Data frame with altitudinal range of species. First
column must be the species name, second column the minimum value of altitude
and the third column the maximum value of altitude}

\item{shp.out}{(logical) Whether the output should be a shapefile as opposed
to a raster.}

\item{resolution}{Numeric value indicating the preferred resolution for the
rasters. Resolution must coarser than the resolution of lc.rec and alt.map.}

\item{continuous}{(logical) Whether the output should be binary or continuous
for the rasters. Only used if resolution provided or when the resolution of
lc.rec and alt.map are different.}

\item{threshold}{Numeric value indicating the threshold of the cell coverage
by the species to the species be considered present (values between 0 and 1).
Only used if continuous = \code{FALSE} and a coarser resolution is provided or
the resolution of lc.rec and alt.map are different or shp.out is
\code{TRUE}.}

\item{extent.out}{Extent object or a vector of four numbers indicating the
preferred output for the rasters. Optional.}

\item{progress}{(logical) a bar showing the progress of the function.}
}
\value{
\code{aoh} returns a list with two elements. The first element is a
data.frame detailing if the function was able (1) or not (0) to refinate the
species distribution. The second element is a list of RasterLayer or
SpatialPolygons object representing the refinated distribution of the species.
For the RasterLayer the value of 0 indicates the cells where the the species
were considerated present before the refinement. Values higher than 0 indicates
the coverage of cell by the spatial distribution after the refinement.
}
\description{
Provide the area of habitat (AOH) of a given species through refinement of its
known geographic distribution
}
\details{
The function map the area of habitat within the geographical
distribution (SpatialPolygon) given as the input data. as the refined
distribution of a given species. This refinement is made considering the
specific preference for habitats of a given species.
}
\examples{
ref_data <- aoh(eoo = sd_amph, lc.rec = lc , matrix.hab.pref = habpref,
alt.map = al, matrix.alt.pref = alpref)
}
\author{
Daniel Gonçalves-Souza & Thaís Dória
}
