% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aohMixS_function.R
\encoding{UTF-8}
\name{aohMixS}
\alias{aohMixS}
\title{aohMixS: mapping the species' AOH by mixing data from EOO and species'
distribution models following Syfert's method.}
\usage{
aohMixS (eooSp = eooSp, modSp = modSp, thresInitial = NULL , thresIncrement = NULL,
continuous = TRUE, poly = NULL, cropToPoly = FALSE, progress = TRUE,
removeTempFile = TRUE, stack = TRUE )
}
\arguments{
\item{eooSp}{Spatial distribution data of the species representing the
original (i.e. not refined) extent of occurrence (EOO). It might correspond to:
\itemize{
  \item path for a folder with spatial distribution files (ESRI shapefile format).
  \item SpatialPolygonsDataFrame (see \code{\link[habitaR]{readShp}} to obtain such
  class of object).
  \item aHull object with elements of 'SpatialPolygons' class created from
  occurrences records (see \code{\link[habitaR]{aHull}} to obtain such class
  of object).
  \item a list of 'RasterLayer' objects created from occurrences records
  (see \code{\link[habitaR]{aHull}} to obtain such class of object).
  \item 'RasterLayer', 'RasterStack' or 'RasterBrick' objects.
 }
NOTE: Input data for EOOs might correspond to Alpha Hull or to Minimum Polygon
Convex (MCP), which can be generated, for example, with the \code{\link[dismo]{convaHull}}
or \code{\link[adehabitatHR]{mcp}} functions.}

\item{modSp}{Species distribution derived from modelling techniques (eg. SDMs).
It might correspond to:
\itemize{
  \item path for a folder with species distribution models files (.asc or .tif
  format).
   \item a list of 'RasterLayer' objects created from occurrences records (see \code{\link[habitaR]{aHull}}
  to obtain such class of object).
  \item 'RasterLayer', 'RasterStack' or 'RasterBrick' objects.
  }}

\item{thresInitial}{The minimum value of threshold to indicate the presence of the
species in the cell. NOTE: From this initial value all threshold spectrum will be
considered up to maximum value of 1.0 based on the increment value defined
by \code{\link[habitaR]{threIncrement}}.}

\item{thresIncrement}{The amount to sequentially increase the initial threshold with each
iteration until find the value that maximizes the geographical overlap (maxJSI) between
EOO and distribution models. See details.}

\item{poly}{Optional. A polygon (ESRI shapefile in a 'SpatialPolygonsDataFrame'
class) of a given area to be used as a mask to crop and restrict the models-derived
EOOs to the area corresponding to provided polygon.}

\item{cropToPoly}{(logical) Whether the output should be cropped by the provided poly.
Only used if 'poly' is provided. If \code{TRUE}, the resulting maps will be restricted
to the area of the poly. Default is \code{FALSE}.}

\item{progress}{(logical) A bar showing the progress of the function.
Default is \code{TRUE}.}

\item{removeTempFile}{(logical) Whether the temporary files generated in each iteration
should be deleted. It is recommended if the user wants to analyse a high volume of data at
the same time. Default is \code{TRUE}.}

\item{stack}{(logical) Whether the output should be returned as a stack of 'RasterLayer'
files. Default is \code{TRUE}.}
}
\value{
By default, \code{aohMix} returns a list with two elements:
 \itemize{
  \item A 'RasterStack' (if \code{stack = TRUE}) or list of 'RasterLayer'
  objects (if \code{stack = FALSE}) of maps representing SDMs-derived EOOs.
   \item A data.frame of species and measures relative to SDM-derived EOO's
   construction and evaluation, as follow:
     \itemize{
        \item Threshold - Threshold value maximizing the geographical similarity
        between distribution from EOO and from model.
        \item MaxJSI - Value of Jaccard Similarity Index (JSI) measured from
        comparison between the EOO and the model returning the maximum geographical
        similarity.
        \item DifSizes - Difference (in number of cells) between the EOO and
        the SDMs-derived EOOs. This measure can be used to evaluate the performance
        of this analysis (see Syfert et al. 2014 for more details).
        \item Prop - Proportion of the EOO maintained in the resulting SDMs-derived EOOs.
        This measure can be used to evaluate the performance of this analysis
        (see Syfert et al. 2014 for more details).
        }
     }
}
\description{
Through a mix approach combining the species' spatial-informed distribution with
modelling techniques, thresholds the distribution models (e.g. SDMs) based on EOO's
geometry to generate models-derived EOOs. Here, EOOs might correspond to the alpha
hulls (ahull) or Minimum Polygon Convex (MCP) and a geographical similarity approach
is used to estimate the shape of the species' distribution based on models. It is
develop by adapting the approach described in Syfert et al. (2014), so that the
species' area of habitat (AOH) can be interpreted as the SDMs-derived EOOs.
}
\details{
Models-derived EOOs are generated through an adapted approach presented by Syfert
et al. (2014) to setting thresholds by maximizing (across all spectrum of
probabilities) the geographical similarity between the EOO (i.e. a polygon drawn
based on the presence records, such as alpha hull or minimum polygon convex) and
the polygon derived from models (e.g. SDMs predicted presences). In each iteration,
based on the increment value defined by \code{\link[habitaR]{threIncrement}}, is
searched the threshold that return the most similar distribution between
the provided model and the EOO through the Jaccard Similarity Index (JSI),
calculated as:
\itemize{
        \item JSI = C / (A + B) - C
        }

A: area of alpha hull / B: area of SDM / C: overlapped area between A and B.
}
\examples{

### Path for folder as Input ###

\dontrun{

# Binary Output #
aohmixR_bin <- aohMixS (eooSp = path_eoo, modSp = path_mod,
thresInitial = 0.05, thresIncrement = 0.25, continuous = FALSE, cropToPoly = TRUE,
poly = poly, progress = TRUE, stack = TRUE)

# Continuous Output #
aohmixR_con <- aohMixS (eooSp = path_eoo, modSp = path_mod,
thresInitial = 0.05, thresIncrement = 0.25, continuous = TRUE, cropToPoly = TRUE,
poly = poly, progress = TRUE, stack = TRUE)

}


### Shapefile as Input ('aHull' or 'SpatialPolygonsDataFrame' class) ###

# Binary Output #

aohmixS_bin <- aohMixS (eooSp = aHull_plantShp, modSp = sdm_plantRas,
thresInitial = 0.05, thresIncrement = 0.25, continuous = FALSE, cropToPoly = TRUE,
poly = poly, progress = TRUE, stack = TRUE)

aohmixS_bin <- aohMixS (eooSp = spdf_plantShp, modSp = sdm_plantRas,
thresInitial = 0.05, thresIncrement = 0.25, continuous = FALSE, cropToPoly = TRUE,
poly = poly, progress = TRUE, stack = TRUE)

# Continuous Output #

aohmixS_con <- aohMixS (eooSp = aHull_plantShp, modSp = sdm_plantRas,
thresInitial = 0.05, thresIncrement = 0.25, continuous = TRUE, cropToPoly = TRUE,
poly = poly, progress = TRUE, stack = TRUE)

aohmixS_con <- aohMixS (eooSp = spdf_plantShp, modSp = sdm_plantRas,
thresInitial = 0.05, thresIncrement = 0.25, continuous = TRUE, cropToPoly = TRUE,
poly = poly, progress = TRUE, stack = TRUE)


### Raster as Input ###

# Binary Output #

aohmixR_bin <- aohMixS (eooSp = aHull_plantRas, modSp = sdm_plantRas,
thresInitial = 0.05, thresIncrement = 0.25, continuous = FALSE, cropToPoly = TRUE,
poly = poly, progress = TRUE, stack = TRUE)

# Continuous Output #

aohmixR_con <- aohMixS (eooSp = aHull_plantRas, modSp = sdm_plantRas,
thresInitial = 0.05, thresIncrement = 0.25, continuous = TRUE, cropToPoly = TRUE,
poly = poly, progress = TRUE, stack = TRUE)

}
\references{
1. Syfert M.M., Joppa L.N., Smith M.J, Coomes D., Bachman S.P. & Brummitt N.A. (2014).
Using species distribution models to inform IUCN Red List assessments. Biological
Conservation. 177.

2. Sangermano F., Eastman J.R. (2012). A GIS framework for the refinement of species
geographic ranges. International Journal of Geographical Information Science 26, 39-55.
}
\author{
Thaís Dória & Daniel Gonçalves-Souza
}
