% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aohVal_function.R
\encoding{UTF-8}
\name{aohVal}
\alias{aohVal}
\title{Validation function}
\usage{
# method for signature 'SpatialPolygonsDataFrame', 'RasterLayer',
'RasterStack', a 'list' of 'RasterLayer' or 'aoh' object (see
 \code{\link[aoh]{aoh}} to generate the latter object)

aohVal(eoo.sp, aoh.sp, plot=TRUE, progress=FALSE)
}
\arguments{
\item{eoo.sp}{Spatial distribution data of the species representing the
original (i.e. not refined) extent of occurrence (EOO). It might correspond to
'SpatialPolygonsDataFrame' (see \code{\link[aoh]{readShp}} to obtain such
class of object), a 'RasterLayer', a 'RasterStack', or a 'list' with 'RasterLayer'
class of features (see \code{\link[aoh]{readRas}} to obtain such object).}

\item{aoh.sp}{Spatial distribution data of the species representing the area
of habitat (AOH). It might correspond to 'SpatialPolygonsDataFrame'
(see \code{\link[aoh]{readShp}} to obtain such class of object), a
'RasterLayer', a 'RasterStack', a 'list' with 'RasterLayer' class of features
(see \code{\link[aoh]{readRas}} to obtain such object), or a 'aoh'
(see \code{\link[aoh]{aoh}} to obtain such class of object).}

\item{plot}{(logical) Whether the output should also return a plot
representing the graphical result of validation (see Rondinini et al. 2011).
Default is \code{TRUE}.}

\item{progress}{(logical) A bar showing the progress of the function.
Default is \code{FALSE}.}
}
\value{
\code{aohVal} returns a data.frame with a summary of results for each species which aoh was evaluated.
It gives the number of records matching with the EOO ("MATCH.EOO) and with the
AOH ("MATCH.AOH) and the values of prevalence of points ("PP", Rondinini et al.
2011) through the proportion of occurrences spatially congruent with suitable
cells, and of model prevalence ("MP", Rondinini et al. 2011), which represents
EOOs proportion assigned as suitable). AOH predict species occurrences
correctly when all available records match with its suitable cells (PP = 1).
The evaluation of AOHs performance is based on difference between PP and MP.
If PP > MP, AOH performs better than EOO in predict presences (Ficetola et al.
2015).

If plot is \code{TRUE}, the function return also a plot with a graphical representation
of validation results (see Rondinini et al. 2011). If plot \code{FALSE}, the function
return a data.frame
}
\description{
Provide the summary of validation results based on comparison between an
independent occurrence records of species and the extent of occurrence (EOO)
and area of habitat (AOH) following Rondinini et al. (2011)
}
\details{
The function gives the summary of validation steps followed to evaluate
the quality of models as described and performed by Rondidini et al. (2011) and
Ficetola et al. (2015).
}
\examples{

### Fictitious amphibians and birds data

data(birds_val)
data(amphs_val)

# Example for signature 'SpatialPolygonsDataFrame' (eoo) and 'aoh' (aoh) with
a 'RasterLayer' class of features.

birdVal1<-aohVal(eoo = eoo_birdShp, aoh = aoh_birdRas, plot = TRUE, progress = TRUE)
amphVal1<-aohVal(eoo = eoo_amphShp, aoh = aoh_amphRas, plot = TRUE, progress = TRUE)


# Example for signature 'RasterLayer' or 'RasterStack' (eoo) and 'aoh' (aoh) with
a 'RasterLayer' class of features.

birdVal2<-aohVal(eoo = eoo_birdRas, aoh = aoh_birdRas, plot = TRUE, progress = TRUE)
amphVal2<-aohVal(eoo = eoo_amphRas, aoh = aoh_amphRas, plot = TRUE, progress = TRUE)


# Example for signature 'SpatialPolygonsDataFrame' (eoo) and 'aoh' (aoh) with
a 'SpatialPolygonsDataFrame' class of features.

birdVal3<-aohVal(eoo = eoo_birdShp, aoh = aoh_birdShp, plot = TRUE, progress = TRUE)
amphVal3<-aohVal(eoo = eoo_amphShp, aoh = aoh_amphShp, plot = TRUE, progress = TRUE)

# Example for signature 'RasterLayer' or 'RasterStack' (eoo) and 'aoh' (aoh) with
a 'SpatialPolygonsDataFrame' class of features.

birdVal4<-aohVal(eoo = eoo_birdRas, aoh = aoh_birdShp, plot = TRUE, progress = TRUE)
amphVal4<-aohVal(eoo = eoo_amphRas, aoh = aoh_amphShp, plot = TRUE, progress = TRUE)

}
\references{
1. Rondinini, C., Di Marco, M., Chiozza, F., Santulli, G., Baisero, D., Visconti,
P., Boitani, L. (2011). Global habitat suitability models of terrestrial mammals.
Philosophical Transactions of the Royal Society B, 366, 2633–2641.

2. Brooks, T. M, Fonseca, S.L. Pimm, Akçakaya, H.R., Buchanan, G.M., …,
Rondinini C. (2019). Measuring Terrestrial Area of Habitat (AOH) and Its
Utility for the IUCN Red List. Trends in Ecology &amp; Evolution, 34(11),
977–986.

3. Ficetola, G. F., Rondinini, C., Bonardi, A., Baisero, D., &amp;
Padoa-Schippa, E. (2015).Habitat availability for amphibians and extinction
threat: a global analysis. Diversity and Distributions, 21(3), 302–311.
}
\author{
Thaís Dória & Daniel Gonçalves-Souza
}
