% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_validation.R
\name{aohVal}
\alias{aohVal}
\title{Validation function}
\usage{
aohVal(eoo, aoh, resolution=NULL, plot=TRUE)
}
\arguments{
\item{eoo}{Spatial distribution data of the species representing the original
(i.e. not refined) extent of occurrence (EOO). It might correspond to
SpatialPolygons in ESRI shapefile (.shp), to Raster file (.asc or .tif) or to a
path for a folder with EOO shapefiles (.shp). If the shapefile is used, the
name of the species must be on the second column of the attribute table
of the shapefile.}

\item{aoh}{Spatial distribution data of the species representing the area
of habitat (AOH). It might correspond to SpatialPolygons in ESRI shapefile (.shp),
to Raster file (.asc or .tif), created or not with function 'aoh', or to
path for a folder with EOO shapefiles (.shp). If the shapefile is used, the
name of the species must be on the second column of the attribute table.}

\item{resolution}{Numeric value to indicate the resolution prefered to generate
the raster files. This argument is only used if the spatial distribution data of
species are in shapefile format. In this case, it should be used the same
resolution adopted to mapping the area of habitat (AOH).}

\item{plot}{(logical) Whether the output should be a plot representing the graphical
result of validation (see Rondinini et al. 2011).}
}
\value{
\code{summary} The result is a data.frame species-specific with a summary
of data. It gives the number of records matching with the EOO ("MATCH.EOO) and with
the AOH ("MATCH.AOH) and the values of prevalence of points ("PP", Rondinini et al. 2011)
through the proportion of occurrences spatially congruent with suitable cells,
and of model prevalence ("MP", Rondinini et al. 2011), which represents EOOs proportion assigned as suitable). AOH predict species occurrences correctly when all available
records match with its suitable cells (PP = 1). The evaluation of quality of
AOHs is based on difference between PP and MP. If PP > MP, AOH performs better than
EOO in predict presences (Ficetola et al. 2015).
}
\description{
Provide the summary of validation results based on comparison between an
independent occurrence records of species and the extent of occurrence (EOO)
and area of habitat (AOH) following Rondinini et al. (2011)
}
\details{
The function give the summary of validation steps followed to evaluate
the quality of models as described and performed by Rondidini et al. (2011)
and Ficetola et al. (2015)
}
\references{
Rondinini, C., Di Marco, M., Chiozza, F., Santulli, G., Baisero, D., Visconti, P., Boitani, L. (2011). Global habitat suitability models of terrestrial mammals. Philosophical Transactions of the Royal Society B, 366, 2633–2641. Brooks, T. M, Fonseca, S.L. Pimm, Akçakaya, H.R., Buchanan, G.M., …, Rondinini C. (2019). Measuring Terrestrial Area of Habitat (AOH) and Its Utility for the IUCN Red List. Trends in Ecology &amp; Evolution, 34(11), 977–986. Ficetola, G. F., Rondinini, C., Bonardi, A., Baisero, D., &amp; Padoa-Schippa, E. (2015).Habitat availability for amphibians and extinction threat: a global analysis. Diversity and Distributions, 21(3), 302–311.
}
\author{
Thaís Dória & Daniel Gonçalves-Souza
}
