% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkOcc_function.R
\encoding{UTF-8}
\name{checkOcc}
\alias{checkOcc}
\title{checkOcc function}
\usage{
checkOcc (poly, sp.occ)
}
\arguments{
\item{poly}{A polygon (ESRI shapefile format) of the specific area to be checked in 
'SpatialPolygonsDataFrame' class.}

\item{sp.occ}{A list of 'SpatialPoints' for multiple species. \code{\link[habitaR]
{readOcc}} can help obtain this data.}
}
\value{
\code{checkOcc} returns a list of species occurring inside of the provided 
polygon with their respective number of ocurrences falling in this polygon. 
total of occurrences inside of the polygon.
}
\description{
From a list of 'SpatialPoints' for multiple species, gives the species and its 
occurrences that are inside of a polygon (ESRI shapefile format) representing 
a region of interest (e.g. study area, continent, country or other regions with 
specific boundaries).
}
\author{
Thaís Dória & Daniel Gonçalves-Souza
}
