% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{class_ref}
\alias{class_ref}
\title{Habitat classification reference}
\format{A data frame with 6 rows and 125 variables}
\usage{
data(class_ref)
}
\description{
A dataset to reference how the user could reclassify the
IUCN's Habitats Classification Scheme for the argument \code{cat}
of the function \code{prefHab}.
}
\keyword{datasets}
