% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pref_hab_function.R
\encoding{UTF-8}
\name{prefHab}
\alias{prefHab}
\title{Habitat preference}
\usage{
prefHab(sp, key = NULL, cat = NULL,
suitability = c('Suitable', 'Marginal', 'Unknown'),
mi = FALSE,
season = c('Resident', 'Breeding','Non-breeding', 'Passage',
'Unknown'), progress = FALSE)
}
\arguments{
\item{sp}{a vector containing the names or IUCN's IDs of the species.}

\item{key}{A IUCN API token. You can request a key in the website
http://apiv3.iucnredlist.org/api/v3/token}

\item{cat}{a two columns data frame. The first column must have the codes
from IUCN's Habitats Classification Scheme (Version 3.1). The second column
must have the numeric code of the user's land use map that is equivalent to
the IUCN's Habitats Classification Scheme. When there is no land use equivalent
to the IUCN's classification, use the value 0. If no data frame is provided the
function uses the dataset obtained through \code{data(class_ref)} as reference.}

\item{suitability}{a vector specifying which levels of habitat suitability
must be considered.}

\item{mi}{(logical) only suitable habitats of major importance must be
included?}

\item{season}{a vector specifying which seasons of habitat importance
must be considered.}

\item{progress}{(logical) a bar showing the progress of the function.}
}
\value{
\code{prefHab} returns a data frame in which the first column is the
name of the species and the posterior columns are the codes of the user's
classification scheme. 1 indicates that the species has preference for a
given class of land use while 0 indicates that it has not.
}
\description{
Obtains the habitat preference of species in IUCN's database and put in the
format required to the function \code{aoh}, based on the classification scheme of
the user's land use map.
}
\examples{
\donttest{
## Not run:
sp <- c('Bradypus torquatus', 'Adelophryne maranguapensis', 'Dermatonotus muelleri')
prefHab(sp)
## End
}

}
\author{
Daniel Gonçalves-Souza & Thaís Dória
}
