% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOcc_function.R
\encoding{UTF-8}
\name{readOcc}
\alias{readOcc}
\title{readOcc: read species occurrences}
\usage{
readOcc (occ = occ, crs = "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0",
distOcc = NULL, occSum = FALSE)
}
\arguments{
\item{occ}{Path for a folder with the species occurrences records files
(.csv format). Coordinates should be in decimal degrees. The names of .csv files
should correspond to the species names and must have 3 columns identified as
"species" (species names or other identification of taxa), "long" (longitude),
"lat" (latitude). NOTE: Longitude must be in a column before the latitude column.}

\item{crs}{The Coordinate Reference System (CRS) specifing the projection and
datum of dataset. Could be a CRS object or a character string.}

\item{distOcc}{A value corresponding to the minimum distance to consider
two coordinates as not duplicate. Values up to this distance will be consider
as duplicates and removed. Units of this value must be in km. Optional.
Default is zero (i.e. only exactly coincident coordinates will be removed).
For more details, see \code{\link[sp:remove.duplicates]{remove.duplicates}}.}

\item{occSum}{(logical) Whether the output should include also a data.frame
with the number of occurrences records after the removal of duplicate coordinates.
Default is \code{FALSE}.}
}
\value{
By default, \code{readOcc} returns a list with two elements:
\itemize{
  \item A 'spOcc' object corresponding to a species list with features of
'SpatialPoints' class.
  \item A data.frame with the number of occurrences records after the cleaning
step to remove duplicate coordinates.
}
If occSum is \code{FALSE}, \code{readOcc} returns only the 'spOcc' object.
}
\description{
Read and clean the occurrences records files (.csv format) of multiples species
from a specified folder and provide a list of 'SpatialPoints' of species through
the removal of duplicates coordinates.
}
\seealso{
\code{\link[sp:remove.duplicates]{remove.duplicates}}
}
\author{
Thaís Dória & Daniel Gonçalves-Souza
}
