% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOcc_function.R
\encoding{UTF-8}
\name{readOcc}
\alias{readOcc}
\title{readOcc function}
\usage{
readCsv (occ, crs, dist)
}
\arguments{
\item{occ}{Path for a folder with the species occurrence records files
(.csv format). The files names should correspond to the species names
and must have be 3 columns identified as "species" (species names or other
identification of taxa), "long" (longitude), "lat" (latitude).}

\item{dist}{A value corresponding to the minimum distance assgined to consider 
two coordinates as not duplicate. Values up to this distance will be considered 
as duplicates and, then, will be removed. Default is 0.0 (in this case, only 
exactly coincindent coordinates will be removed). Units of this values must be 
the same as those of the coordinates for projected data or in km if coordinates 
are defined to be longitude/latitude. For more details, see "remove.duplicates" 
function.}

\item{crs}{The Coordinate Reference System (CRS) specifing the projection and 
datum of dataset. Could be a CRS object or a character string}
}
\value{
\code{readOcc} returns a 'sp.occ' object corresponding to a list of species
with features from a class of 'SpatialPoints'.
}
\description{
Read and clean the occurrences records files (.csv format) of multiples species
from a specified folder and provide a list of 'SpatialPoints' of species through 
the removal of duplicates coordinates.
}
\author{
Thaís Dória & Daniel Gonçalves-Souza
}
