% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRas_funciton.R, R/readRas_function.R
\name{readRas}
\alias{readRas}
\title{readRAS function}
\usage{
readRas (path)

readRas (path)
}
\arguments{
\item{path}{Path for a folder with spatial distribution raster files
(.asc or .tif format).}

\item{path}{Path for a folder with spatial distribution raster files
(.asc or .tif format).}
}
\value{
A list of raster files

A list of raster files or a raster file
}
\description{
Read the spatial distribution raster files (.tif or .asc format) of species
from a specified folder and provide a list with the raster files.

Read the spatial distribution raster files (.tif or .asc format) of species
from a specified folder and provide a list with the raster files.
}
\author{
Thaís Dória & Daniel Gonçalves-Souza

Thaís Dória & Daniel Gonçalves-Souza
}
